refactor : extract getTrackerBundleReportMode
